fixup! rename sentry setup file
