+++ main/setup-sentry.js
